feat: move all DTOs to competition.dto package as internal DTOs
