+++ src/main/java/com/itasocialacademy/oitassist/competition/dto/response/CompetitionResponse.java
@@ -1,4 +1,4 @@
-package com.itasocialacademy.oitassist.competition.dao.dto.response;
+package com.itasocialacademy.oitassist.competition.dto.response;
 
 import com.itasocialacademy.oitassist.competition.dao.enums.CompetitionStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
