+++ src/main/java/com/itasocialacademy/oitassist/competition/dto/response/StageTreeResponse.java
@@ -1,0 +1,10 @@
+package com.itasocialacademy.oitassist.competition.dto.response;
+
+import java.util.List;
+import lombok.Builder;
+
+@Builder
+public record StageTreeResponse(
+    StageResponse stage,
+    List<TourResponse> tours) {
+}
