+++ src/main/java/com/itasocialacademy/oitassist/competition/dto/response/TourResponse.java
@@ -1,4 +1,4 @@
-package com.itasocialacademy.oitassist.competition.dao.dto.response;
+package com.itasocialacademy.oitassist.competition.dto.response;
 
 import java.time.ZonedDateTime;
 import lombok.Builder;
