+++ src/main/java/com/itasocialacademy/oitassist/competition/dto/validation/HasDateRange.java
@@ -1,4 +1,4 @@
-package com.itasocialacademy.oitassist.competition.dao.dto.validation;
+package com.itasocialacademy.oitassist.competition.dto.validation;
 
 import com.itasocialacademy.oitassist.competition.validation.DateRangeValidator;
 import java.time.ZonedDateTime;
