+++ src/main/java/com/itasocialacademy/oitassist/competition/mapper/CompetitionMapper.java
@@ -1,7 +1,7 @@
 package com.itasocialacademy.oitassist.competition.mapper;
 
-import com.itasocialacademy.oitassist.competition.dao.dto.request.CreateCompetitionRequest;
-import com.itasocialacademy.oitassist.competition.dao.dto.response.CompetitionResponse;
+import com.itasocialacademy.oitassist.competition.dto.request.CreateCompetitionRequest;
+import com.itasocialacademy.oitassist.competition.dto.response.CompetitionResponse;
 import com.itasocialacademy.oitassist.competition.dao.model.Competition;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
