+++ src/main/java/com/itasocialacademy/oitassist/competition/mapper/StageMapper.java
@@ -1,7 +1,7 @@
 package com.itasocialacademy.oitassist.competition.mapper;
 
-import com.itasocialacademy.oitassist.competition.dao.dto.request.CreateStageRequest;
-import com.itasocialacademy.oitassist.competition.dao.dto.response.StageResponse;
+import com.itasocialacademy.oitassist.competition.dto.request.CreateStageRequest;
+import com.itasocialacademy.oitassist.competition.dto.response.StageResponse;
 import com.itasocialacademy.oitassist.competition.dao.model.Stage;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
