+++ src/main/java/com/itasocialacademy/oitassist/competition/service/CompetitionServiceImpl.java
@@ -1,10 +1,10 @@
 package com.itasocialacademy.oitassist.competition.service;
 
-import com.itasocialacademy.oitassist.competition.api.dto.CompetitionTreeResponse;
-import com.itasocialacademy.oitassist.competition.api.dto.StageTreeResponse;
-import com.itasocialacademy.oitassist.competition.dao.dto.request.CreateCompetitionRequest;
-import com.itasocialacademy.oitassist.competition.dao.dto.response.CompetitionResponse;
-import com.itasocialacademy.oitassist.competition.dao.dto.response.TourResponse;
+import com.itasocialacademy.oitassist.competition.dto.response.CompetitionTreeResponse;
+import com.itasocialacademy.oitassist.competition.dto.response.StageTreeResponse;
+import com.itasocialacademy.oitassist.competition.dto.request.CreateCompetitionRequest;
+import com.itasocialacademy.oitassist.competition.dto.response.CompetitionResponse;
+import com.itasocialacademy.oitassist.competition.dto.response.TourResponse;
 import com.itasocialacademy.oitassist.competition.dao.enums.CompetitionStatus;
 import com.itasocialacademy.oitassist.competition.dao.model.Competition;
 import com.itasocialacademy.oitassist.competition.dao.model.Stage;
