+++ src/main/java/com/itasocialacademy/oitassist/competition/service/StageServiceImpl.java
@@ -1,8 +1,8 @@
 package com.itasocialacademy.oitassist.competition.service;
 
-import com.itasocialacademy.oitassist.competition.dao.dto.request.CreateStageRequest;
-import com.itasocialacademy.oitassist.competition.dao.dto.request.UpdateStageRequest;
-import com.itasocialacademy.oitassist.competition.dao.dto.response.StageResponse;
+import com.itasocialacademy.oitassist.competition.dto.request.CreateStageRequest;
+import com.itasocialacademy.oitassist.competition.dto.request.UpdateStageRequest;
+import com.itasocialacademy.oitassist.competition.dto.response.StageResponse;
 import com.itasocialacademy.oitassist.competition.dao.model.Stage;
 import com.itasocialacademy.oitassist.competition.dao.repository.StageRepository;
 import com.itasocialacademy.oitassist.competition.exceptions.CompetitionHierarchyValidationException;
