+++ src/main/java/com/itasocialacademy/oitassist/competition/validation/DateRangeValidator.java
@@ -1,7 +1,7 @@
 package com.itasocialacademy.oitassist.competition.validation;
 
-import com.itasocialacademy.oitassist.competition.dao.dto.validation.HasDateRange;
-import com.itasocialacademy.oitassist.competition.dao.dto.validation.ValidDateRange;
+import com.itasocialacademy.oitassist.competition.dto.validation.HasDateRange;
+import com.itasocialacademy.oitassist.competition.dto.validation.ValidDateRange;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
